+++ src/main/java/entityparsers/ParseCarBuyerUseCase.java
@@ -6,6 +6,8 @@ import entities.CarBuyer;
 
 import entitybuilder.GenerateBuyerUseCase;
 
+import javax.json.JsonObject;
+
 public class ParseCarBuyerUseCase {
 
     private final AttributeMap map;
@@ -27,7 +29,7 @@ public class ParseCarBuyerUseCase {
     public CarBuyer parse() throws ClassCastException {
         GenerateBuyerUseCase buyerGenerator = new GenerateBuyerUseCase();
         AttributeMap buyerMap = (AttributeMap) map.getItem(EntityStringNames.BUYER_STRING);
-        int budget = (int) buyerMap.getItem("pytBudget").getAttribute();
+        double budget = (double) buyerMap.getItem("pytBudget").getAttribute();
         int creditScore = (int) buyerMap.getItem("creditScore").getAttribute();
         return buyerGenerator.GenerateBuyerDataUseCase(budget, creditScore);
     }
